Simplify shader injection for lights with a temporary workaround;
The shader source for lights currently relies on injecting a source which itself contains an injection tag.
This creates an implicit order-dependency in the injection process; we should remove it if possible. For now
there is a workaround: injections occur in alphabetical order of the tag name, so dependency can be encoded
in the tag names.