Fixed invlid future shared state issue
Its possible with fast zoomin-out that a thread starts processing a tile which has been deleted from the visible set.
In such cases, the future of that thread should return a nullptr for that tile.

Previously, we were seeing a invalid future state in such a case. This commit fixes this behavior.