Fix input using modifier keys on Linux
When pressing shift or control on Ubuntu, GLFW receives a key event with the shift modifier inactive on the down event and active on the up event.
This is opposite to the behavior in OS X, so to fix this we now just poll both shift and control keys during the scroll callback.