Rotating the map is now mapped to alt+scroll
Shift+scroll was an issue because on OS X, holding shift will map scroll wheel input to the x axis (but not touchpad scrolling). I tried just taking the input axis with the largest magnitude, but it created jittery behavior with touchpad scrolling so it seems best to just use a different key.