disable LoDDiscard for now:
Labels of tiles are only updated when the tile passes the
LoD discard test. This means that a tile that passed the test
once but is discarded afterwards still draws its label meshes
without being updated.

One solution (or rather workaround for now) could be to use
an 'active' flag for the VboMesh.