Multiple improvements
Using Ekta for user testing and suggestion

- Much better distinction in shove and scale/rotate gestures.
- Mimics google and apple maps behavior of how rotation is delayed to allow accidental scale hiccups and also better
detect shove gesture.
- Shove gesture is reversed to mimic how its done in google and apple maps.

- TODO:
a. Rotate about 2nd finger/stationary finger vs rotate about focus point.
b. Make Tilt about the center of the view. (which is basically bringing the camera down and back) ... Google and Apple
maps behave the same way.
c. Deciding on a reasonable tilt angle (when to stop).
d. Scale behavior/misbehavior when tilted!!! (Thats because we want to keep the localation of the points underneeth the
finger pointers the same with motion).