Better  Tilting
Concepts
- counter rotate the "at" point about x axis. (previously done)
- counter rotate "up" vector about -y axis. (previously done)

But since the field of view of camera remains the same we will get clipping and old view origin will be left behind.
To handle this we translate the view by the difference of new origin and the old origin (in world space).
Also making sure this only done when a pitch is changed ie a tilt/shove gesture was performed (during this View::update
call).