Update Texture data when network task finishes for a texture
- While parsing the scene file if the a texture is in process of being downloaded, a dummy black 1
pixel texture is created as a placeholder. When the network task finished with the appropriate
texture data, this placeholder Texture object's data is updated with the downloaded data. If this
texture is a spriteAtlas, respective spriteNodes are updated with the new data of this texture.

- Adds some helper methods on spriteAtlas to update sprite Nodes with updated texture data

- Handle absolute texture paths
    - Should be updated aptly when a generalized URI handler is implemented

TODO: Second scene's loading process should wait untill current scene has finished loading. Make
sure only the latest scene is sent for loading (all in between scenes which have not started loading
should be ignored)