Fix global property updates
since global properties are substituded when parsing the yaml scene for the first time, there was no information of
where the global properties were applied when any of these globals were updated.

This fixes the issue, by keeping track of the path of the node(s) which were using the global properties to begin with
so that subsequent global property updates are applied.

Handle scene updates more robustly

make sure previous global can be explicitly set to other values, and can be reset to another global!