Scene sets the pixel scale of its Styles
The Scene 'owns' its Styles, so when setting a pixel scale for the Scene, it should be applied to Styles in the Scene too.

This fixes a previous inconsistency: the correct pixel scale was set on the Scene in Map::Impl::setScene, but the Styles in the Scene would still have their default pixel scale.