Fix zip bundled resource path
This fixes a bug where an assets path within a zip bundle was wrong. (relative url of an asset with
respect to its base scene is not same as the path of the asset within a zip bundle).

Now a zip resource's bundled path is determined based on its absolute path and absolute path of the
zip bundle itself.