Fix possible dereference of IOSPlatform instance after deletion
Capturing 'm_mapView' implicitly captures 'this' and dereferences it in the block.

If a TGURLHandler outlives its TGMapView it can invoke this callback after the IOSPlatform is deleted, causing an access error.

With this change we can safely check whether the TGMapView is deleted and exit the block early if so.