Fix PolygonBuilder name collision (#2054)
Rename ClientGeoJsonSource to ClientDataSource

Nest PolygonBuilder and PolylineBuilder in ClientDataSource

PolygonBuilder collided with an existing struct with the same fully-qualified name in core/src/util/builders.h
This resulted in undefined behavior. With clang, the linker chose destructors incorrectly for one of the instances, leading to a segfault when the Map is destructed.
Nesting the newer PolygonBuilder within ClientDataSource resolves the collision.